+++ rust/main/hyperlane-base/src/settings/chains.rs
@@ -514,7 +514,10 @@ impl ChainConf {
                     conf,
                     &locator,
                     metrics,
-                    h_midl::SequenceIndexerBuilder { reorg_period },
+                    h_midl::SequenceIndexerBuilder {
+                        reorg_period,
+                        finality: conf.finality.clone(),
+                    },
                 )
                 .await
             }
@@ -607,7 +610,10 @@ impl ChainConf {
                     conf,
                     &locator,
                     metrics,
-                    h_midl::DeliveryIndexerBuilder { reorg_period },
+                    h_midl::DeliveryIndexerBuilder {
+                        reorg_period,
+                        finality: conf.finality.clone(),
+                    },
                 )
                 .await
             }
@@ -781,6 +787,7 @@ impl ChainConf {
                     h_midl::InterchainGasPaymasterIndexerBuilder {
                         mailbox_address: self.addresses.mailbox.into(),
                         reorg_period,
+                        finality: conf.finality.clone(),
                     },
                 )
                 .await
@@ -863,7 +870,10 @@ impl ChainConf {
                     conf,
                     &locator,
                     metrics,
-                    h_midl::MerkleTreeHookIndexerBuilder { reorg_period },
+                    h_midl::MerkleTreeHookIndexerBuilder {
+                        reorg_period,
+                        finality: conf.finality.clone(),
+                    },
                 )
                 .await
             }